+++ crates/jarl-core/src/lints/base/lengths/lengths.rs
@@ -77,13 +77,14 @@ pub fn lengths(ast: &RCall, fn_name: &str) -> anyhow::Result<Option<Diagnostic>>
             .text_trimmed()
             == "length"
     {
+        let arg_x = unwrap_or_return_none!(arg_x.and_then(|arg| arg.value()));
         let range = ast.syntax().text_trimmed_range();
         let diagnostic = Diagnostic::new(
             Lengths,
             range,
             Fix::new(
                 range,
-                format!("lengths({})", arg_x.unwrap().into_syntax().text_trimmed()),
+                format!("lengths({})", arg_x.into_syntax().text_trimmed()),
                 node_contains_comments(ast.syntax()),
             ),
         );