+++ crates/jarl/tests/integration/rules.rs
@@ -636,6 +636,7 @@ fn test_select_all_keyword() -> anyhow::Result<()> {
     let case = CliTest::with_file(
         "test.R",
         "
+library(testthat)
 any(is.na(x))
 expect_equal(foo(x), TRUE)
 ",
@@ -656,17 +657,17 @@ expect_equal(foo(x), TRUE)
     exit_code: 1
     ----- stdout -----
     warning: any_is_na
-     --> test.R:2:1
+     --> test.R:3:1
       |
-    2 | any(is.na(x))
+    3 | any(is.na(x))
       | ------------- `any(is.na(...))` is inefficient.
       |
       = help: Use `anyNA(...)` instead.
 
     warning: expect_true_false
-     --> test.R:3:1
+     --> test.R:4:1
       |
-    3 | expect_equal(foo(x), TRUE)
+    4 | expect_equal(foo(x), TRUE)
       | -------------------------- `expect_equal(x, TRUE)` is not as clear as `expect_true(x)`.
       |
       = help: Use `expect_true(x)` instead.
@@ -697,9 +698,9 @@ expect_equal(foo(x), TRUE)
     exit_code: 1
     ----- stdout -----
     warning: any_is_na
-     --> test.R:2:1
+     --> test.R:3:1
       |
-    2 | any(is.na(x))
+    3 | any(is.na(x))
       | ------------- `any(is.na(...))` is inefficient.
       |
       = help: Use `anyNA(...)` instead.
@@ -721,6 +722,7 @@ fn test_extend_select() -> anyhow::Result<()> {
     let case = CliTest::with_file(
         "test.R",
         "
+library(testthat)
 any(is.na(x))
 expect_equal(foo(x), TRUE)
 ",
@@ -741,17 +743,17 @@ expect_equal(foo(x), TRUE)
     exit_code: 1
     ----- stdout -----
     warning: any_is_na
-     --> test.R:2:1
+     --> test.R:3:1
       |
-    2 | any(is.na(x))
+    3 | any(is.na(x))
       | ------------- `any(is.na(...))` is inefficient.
       |
       = help: Use `anyNA(...)` instead.
 
     warning: expect_true_false
-     --> test.R:3:1
+     --> test.R:4:1
       |
-    3 | expect_equal(foo(x), TRUE)
+    4 | expect_equal(foo(x), TRUE)
       | -------------------------- `expect_equal(x, TRUE)` is not as clear as `expect_true(x)`.
       |
       = help: Use `expect_true(x)` instead.
@@ -780,17 +782,17 @@ expect_equal(foo(x), TRUE)
     exit_code: 1
     ----- stdout -----
     warning: any_is_na
-     --> test.R:2:1
+     --> test.R:3:1
       |
-    2 | any(is.na(x))
+    3 | any(is.na(x))
       | ------------- `any(is.na(...))` is inefficient.
       |
       = help: Use `anyNA(...)` instead.
 
     warning: expect_true_false
-     --> test.R:3:1
+     --> test.R:4:1
       |
-    3 | expect_equal(foo(x), TRUE)
+    4 | expect_equal(foo(x), TRUE)
       | -------------------------- `expect_equal(x, TRUE)` is not as clear as `expect_true(x)`.
       |
       = help: Use `expect_true(x)` instead.